pyproject.toml
[project]
name = "researchmind"
version = "0.1.0"
description = "AI-powered research assistant"
authors = [
    {name = "Dhara Shah",email = "dharadshah@gmail.com"}
]
python = ">=3.13,<3.14"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "gradio (>=6.13.0,<7.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "langchain-openai (>=1.2.1,<2.0.0)",
    "langchain-ollama (>=1.1.0,<2.0.0)",
    "langchain-core (>=1.3.2,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "sentence-transformers (>=5.4.1,<6.0.0)",
    "chromadb (>=1.5.8,<2.0.0)",
    "qdrant-client (>=1.17.1,<2.0.0)",
    "llama-index-readers-web (>=0.6.0,<0.7.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "arxiv (>=3.0.0,<4.0.0)",
    "ddgs (>=9.14.1,<10.0.0)",
    "langgraph (>=1.1.9,<2.0.0)",
    "langsmith (>=0.7.37,<0.8.0)",
    "trulens-eval (>=2.7.2,<3.0.0)",
    "presidio-analyzer (>=2.2.362,<3.0.0)",
    "presidio-anonymizer (>=2.2.362,<3.0.0)",
    "spacy (>=3.8.14,<4.0.0)",
    "langgraph-checkpoint-sqlite (>=3.0.3,<4.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
